requirements.txt
absl-py==2.0.0
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dm-tree==0.1.8
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.46.0
fpdf2==2.7.7
gast==0.5.4
google-auth==2.25.1
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
h5py==3.10.0
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keras==2.15.0
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.2.0
more-itertools==8.10.0
namex==0.0.7
netifaces==0.11.0
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==3.1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==5.4.1
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scipy==1.11.4
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow-cpu==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.1.0
uWSGI==2.0.23
wadllib==1.3.6
waitress==2.1.2
Werkzeug==3.0.1
wrapt==1.14.1
zipp==1.0.0